extv234 support newer software version
Newer versions of the software on the kobo uses blkid (provided by
busybox) to check whether the partition to export through usb is vfat.
This patch overshadows the systems blkid with a version which cheekily
lies about any extX-partition named "KOBOeReader" and tells the system
it is vfat and should therefore be umounted. YMMV.